refactor: 패스워드 커스텀 훅 삭제
