--- src/entities/card/password.ts
@@ -1,10 +1,0 @@
-export const PASSWORD_LENGTH = 2;
-
-const ERROR_MESSAGE = {
-  RANGE: 'PASSWORD를 전부 채워주세요.',
-};
-
-export const validatePassword = (password: string): string | undefined => {
-  if (password.length !== PASSWORD_LENGTH) return ERROR_MESSAGE.RANGE;
-  return;
-};
